Section PUT and DELETE methods' response changed * Section PUT and DELETE methods are cast to single response if section was not found

+++ src/main/java/org/fukua/demo/service/SectionService.java
@@ -48,6 +48,9 @@ public class SectionService implements Validator{
     }
 
     public void deleteSection(long id) {
+        if (sectionRepository.getById(id) == null) {
+            throw new SectionIsNotFoundException();
+        }
         sectionRepository.deleteById(id);
     }
